+++ src/main/java/com/soda/project/stats/repository/ProjectDailyStatsRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.soda.project.repository;
+package com.soda.project.stats.repository;
 
 import com.querydsl.core.Tuple;
 import com.soda.project.dto.ProjectStatsCondition;
